chore(PHASE-3): feat(interview-room): integrate raw audio capture and transmission PHASE-3-101
Introduces functionality for raw PCM audio capture via `AudioRecorderService` and sends base64 encoded audio data through the WebSocket client. The speech recognition setup is replaced with the new audio recording logic, enhancing real-time audio processing. Updated the audio data callback to utilize `Int16Array` for improved compatibility.

+++ frontend/src/services/AudioRecorderService.ts
@@ -23,7 +23,7 @@ export class AudioRecorderService {
   private workletNode: AudioWorkletNode | null = null;
   private recording: boolean = false;
   private config: AudioRecorderConfig;
-  private dataCallback: ((data: Float32Array) => void) | null = null;
+  private dataCallback: ((data: Int16Array) => void) | null = null;
 
   constructor(config: AudioRecorderConfig = {}) {
     this.config = {
@@ -126,7 +126,7 @@ export class AudioRecorderService {
   /**
    * Register callback for PCM data
    */
-  onData(callback: (data: Float32Array) => void): void {
+  onData(callback: (data: Int16Array) => void): void {
     this.dataCallback = callback;
   }
